# docs/contracts/assistant-v2.openapi.yaml
openapi: 3.0.3
info:
  title: Linguacast Research Assistant API v2
  version: 2.0.0
  description: >-
    V15 file-workspace research-assistant contract. Clients create a Research,
    stream durable turn events, read artifacts by ID, confirm transcription
    and global memory proposals. This document, assistant-v2.wire.schema.json,
    assistant-v2-error-codes.md, assistant-sse-v2.md and
    assistant-workspace-manifest-v1.schema.json are the single source of truth.
    V1 contracts remain frozen and must not be reused by name.
  contact:
    name: PodcastEnglishStudio V15
servers:
  - url: https://assistant.linguacast.xyz
    description: Production entry (proxy routes /v2/assistant* to 127.0.0.1:3230)
...
security:
  - bearerAuth: []
paths:
  /v2/assistant/researches:
    post:
      tags: [researches]
      summary: Create a Research, workspace, initial manifest, and grants
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchCreateRequest'
      responses:
        '201':
          description: Database row, directory, and initial manifest are committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Research'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
    get:
      tags: [researches]
      summary: List V2 Research summaries newest first. Never includes V1 sessions.
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - $ref: '#/components/parameters/ClientVersion'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Page of Research summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ClientVersion:
      name: X-Client-Version
      in: header
      required: false
      schema:
        type: string
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Illegal state, idempotency conflict, or expired SSE cursor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: Saturated, V2 disabled, or dependency not ready
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ResearchId:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    TurnId:
      type: string
      pattern: '^vt_[0-9A-HJKMNP-TV-Z]{26}$'
...
    GrantAlias:
      type: string
      pattern: '^[a-z][a-z0-9-]{0,62}$'
...
    TranslationQuality:
      type: string
      enum: [fast, quality]
...
    GrantPermission:
      type: string
      enum: [read, read_write]
...
    ErrorBody:
      type: object
      required: [code, message, retryable, traceId]
      additionalProperties: true
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
        retryable:
          type: boolean
        retryAfterSeconds:
          type: integer
          minimum: 0
        traceId:
          type: string
          minLength: 1
        params:
          type: object
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    ResearchCreateRequest:
      type: object
      additionalProperties: false
      properties:
        outputLanguage:
          type: string
          minLength: 2
        storefront:
          type: string
          minLength: 2
          maxLength: 2
        targetLanguage:
          type: string
          minLength: 2
        translationQuality:
          $ref: '#/components/schemas/TranslationQuality'
        title:
          type: string
          maxLength: 80
        sharedAliases:
          type: array
          maxItems: 16
          items:
            $ref: '#/components/schemas/GrantAlias'
...
    WorkspaceGrant:
      type: object
      required: [alias, permission, allowedExtensions, maxFileBytes, grantedAt]
      additionalProperties: true
      description: Never includes a real server path.
      properties:
        alias:
          $ref: '#/components/schemas/GrantAlias'
        permission:
          $ref: '#/components/schemas/GrantPermission'
        allowedExtensions:
          type: array
          items:
            type: string
        maxFileBytes:
          type: integer
          minimum: 1
        grantedAt:
          type: string
          format: date-time
        status:
          type: string
          description: ready | unavailable | unknown
...
    ArtifactCounts:
      type: object
      additionalProperties:
        type: integer
        minimum: 0
      properties:
        web_search:
          type: integer
          minimum: 0
        web_page:
          type: integer
          minimum: 0
        podcast_search:
          type: integer
          minimum: 0
        youtube_search:
          type: integer
          minimum: 0
        transcript:
          type: integer
          minimum: 0
        research_memory:
          type: integer
          minimum: 0
        report:
          type: integer
          minimum: 0
...
    Research:
      type: object
      required: [researchId, title, status, workspaceStatus, createdAt, updatedAt]
      additionalProperties: true
      properties:
        researchId:
          $ref: '#/components/schemas/ResearchId'
        title:
          type: string
        phase:
          type: string
        status:
          type: string
          description: creating | ready | deleting | deleted | failed | degraded | corrupt | unknown
        workspaceStatus:
          type: string
        outputLanguage:
          type: string
        storefront:
          type: string
        targetLanguage:
          type: string
        translationQuality:
          $ref: '#/components/schemas/TranslationQuality'
        activeTurnId:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/TurnId'
        artifactCounts:
          $ref: '#/components/schemas/ArtifactCounts'
        grants:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceGrant'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResearchListResponse:
      type: object
      required: [researches, nextCursor]
      additionalProperties: false
      properties:
        researches:
          type: array
          items:
            $ref: '#/components/schemas/Research'
        nextCursor:
          nullable: true
          type: string
...